\name{Tunisian Cities by Zip Code}
\alias{Tunisia Zip Code}
\title{Tunisia Zip Code}
\usage{
Call the function TunisiaZipCode().

Type a valid zip code and submit to search for Tunisian
cities, delegations and governorates.
A valid zip code is contained in this table:

\tabular{ll}{
Gouvernorate\tab Zip Code\cr
Gouvernorate of l'Ariana    \tab 20**           \cr
Gouvernorate of Béja        \tab 90**           \cr
Gouvernorate of Ben Arous   \tab 20** and  11** \cr
Gouvernorate of Bizerte     \tab 70**           \cr
Gouvernorate of Gabès       \tab 60**           \cr
Gouvernorate of Gafsa       \tab 21**           \cr
Gouvernorate of Jendouba    \tab 81**           \cr
Gouvernorate of Kairouan    \tab 31**           \cr
Gouvernorate of Kasserine   \tab 12**           \cr
Gouvernorate of Kébili      \tab 42**           \cr
Gouvernorate of la Manouba  \tab 20** and 11**  \cr
Gouvernorat du Kef          \tab 71**           \cr
Gouvernorate of Mahdia      \tab 51**           \cr
Gouvernorate of Médenine    \tab 41**           \cr
Gouvernorate of Monastir    \tab 50**           \cr
Gouvernorate of Nabeul      \tab 80**           \cr
Gouvernorate of Sfax        \tab 30**           \cr
Gouvernorate of Sidi Bouzid \tab 91**           \cr
Gouvernorate of Siliana     \tab 61**           \cr
Gouvernorate of Sousse      \tab 40**           \cr
Gouvernorate of Tataouine   \tab 32**           \cr
Gouvernorate of Tozeur      \tab 22*            \cr
Gouvernorate of Tunis       \tab 20** and 10**  \cr
Gouvernorate of Zaghouan    \tab 11**           \cr
}

